Improve the body of the for loop
I was using the std::vector.front method to access the elements
in the vector: When I'm done with some element, I've deleted it so
that I can still use std::vector.front method to access next element.
But this has a very big problem behind it that I didn't realize while
I'm writing the code. The delete operation was not O(1) since I'm not
deleting from the back. It was shifting all the values after the 0th
element which causes too much time to unnecessarily spent. So I
removed that part and rewrited the logic. Now the algorithm runs in
O(nlogn) as it is supposed to be.